GbasetAdapter

+++ anyline-data-jdbc-dialect/anyline-data-jdbc-gbase/src/main/java/org/anyline/data/jdbc/gbase/GbasetAdapter.java
@@ -43,7 +43,7 @@ public class GbasetAdapter extends SQLAdapter implements JDBCAdapter, Initializi
 		}
 		PageNavi navi = run.getPageNavi();
 		if(null != navi){
-			int limit = navi.getLastRow() - navi.getFirstRow() + 1;
+			long limit = navi.getLastRow() - navi.getFirstRow() + 1;
 			if(limit < 0){
 				limit = 0;
 			}
